requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-recaptcha==4.0.0
django-rosetta==0.10.0
django-tinymce==4.1.0
idna==3.7
pillow==10.4.0
polib==1.2.0
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
